weather tutorial and documentation enhancements
