+++ docs/flutterweathertutorial.md
@@ -530,7 +530,7 @@ class WeatherRepository {
 
   Future<Weather> getWeather(String city) async {
     final int locationId = await weatherApiClient.getLocationId(city);
-    return await weatherApiClient.fetchWeather(locationId);
+    return weatherApiClient.fetchWeather(locationId);
   }
 }
 ```
@@ -687,23 +687,37 @@ That's all there is to it! Now we're ready to move on to the final layer: the pr
 
 ### Setup
 
-As you've probably already seen in other tutorials, we're going to create a `SimpleBlocDelegate` so that we can see all state transitions in our application. Let's go ahead and add this to our `main.dart` file above the `main()` function. Then go ahead and delete all the code below `main`
+As you've probably already seen in other tutorials, we're going to create a `SimpleBlocDelegate` so that we can see all state transitions in our application. Let's go ahead and create `simple_bloc_delegate.dart` and create our own custom delegate.
 
 ```dart
 import 'package:bloc/bloc.dart';
 
 class SimpleBlocDelegate extends BlocDelegate {
   @override
+  void onEvent(Bloc bloc, Object event) {
+    super.onEvent(bloc, event);
+    print('onEvent $event');
+  }
+
+  @override
   onTransition(Bloc bloc, Transition transition) {
     super.onTransition(bloc, transition);
-    print(transition);
+    print('onTransition $transition');
+  }
+
+  @override
+  void onError(Bloc bloc, Object error, StackTrace stacktrace) {
+    super.onError(bloc, error, stacktrace);
+    print('onError $error');
   }
 }
 ```
 
-Next, we're going to set our delegate in our `main` function like so:
+We can then import it into `main.dart` file and set our delegate like so:
 
 ```dart
+import 'package:flutter_weather/simple_bloc_delegate.dart';
+
 void main() {
   BlocSupervisor.delegate = SimpleBlocDelegate();
   runApp(App());
@@ -1204,6 +1218,38 @@ if (event is RefreshWeather) {
 ```
 
 Here we are just creating a new event that will ask our weatherRepository to make an API call to get the weather for the city.
+
+We can refactor `mapEventToState` to use some private helper functions in order to keep the code organized and easy to follow:
+
+```dart
+@override
+Stream<WeatherState> mapEventToState(WeatherEvent event) async* {
+  if (event is FetchWeather) {
+    yield* _mapFetchWeatherToState(event);
+  } else if (event is RefreshWeather) {
+    yield* _mapRefreshWeatherToState(event);
+  }
+}
+
+Stream<WeatherState> _mapFetchWeatherToState(FetchWeather event) async* {
+  yield WeatherLoading();
+  try {
+    final Weather weather = await weatherRepository.getWeather(event.city);
+    yield WeatherLoaded(weather: weather);
+  } catch (_) {
+    yield WeatherError();
+  }
+}
+
+Stream<WeatherState> _mapRefreshWeatherToState(RefreshWeather event) async* {
+  try {
+    final Weather weather = await weatherRepository.getWeather(event.city);
+    yield WeatherLoaded(weather: weather);
+  } catch (_) {
+    yield state;
+  }
+}
+```
 
 Lastly, we need to update our presentation layer to use a `RefreshIndicator` widget. Let's go ahead and modify our `Weather` widget in `widgets/weather.dart`. There are a few things we need to do.
