+++ examples/flutter_weather/lib/blocs/weather_bloc.dart
@@ -63,22 +63,28 @@ class WeatherBloc extends Bloc<WeatherEvent, WeatherState> {
   @override
   Stream<WeatherState> mapEventToState(WeatherEvent event) async* {
     if (event is FetchWeather) {
-      yield WeatherLoading();
-      try {
-        final Weather weather = await weatherRepository.getWeather(event.city);
-        yield WeatherLoaded(weather: weather);
-      } catch (_) {
-        yield WeatherError();
-      }
+      yield* _mapFetchWeatherToState(event);
+    } else if (event is RefreshWeather) {
+      yield* _mapRefreshWeatherToState(event);
     }
+  }
 
-    if (event is RefreshWeather) {
-      try {
-        final Weather weather = await weatherRepository.getWeather(event.city);
-        yield WeatherLoaded(weather: weather);
-      } catch (_) {
-        yield state;
-      }
+  Stream<WeatherState> _mapFetchWeatherToState(FetchWeather event) async* {
+    yield WeatherLoading();
+    try {
+      final Weather weather = await weatherRepository.getWeather(event.city);
+      yield WeatherLoaded(weather: weather);
+    } catch (_) {
+      yield WeatherError();
+    }
+  }
+
+  Stream<WeatherState> _mapRefreshWeatherToState(RefreshWeather event) async* {
+    try {
+      final Weather weather = await weatherRepository.getWeather(event.city);
+      yield WeatherLoaded(weather: weather);
+    } catch (_) {
+      yield state;
     }
   }
 }
